+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # Changelog
 
+## 1.1.0 - 2016-10-17
+
+Added Google Reverse method
+
+```javascript
+geocoder.googleReverse([-75.1, 45.1])
+```
+
 ## 1.0.0 - 2016-10-14
 
 First Stable release was created